third_party/WebKit/LayoutTests/css-parser/text-indent-valid.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/property-parsing-test.js"></script>
<script>

assert_valid_value("text-indent", "each-line calc(2em + 3ex)", "calc(2em + 3ex) each-line");
assert_valid_value("text-indent", "each-line hanging -5vh", "-5vh hanging each-line");
</script>
